normalize vec_string
