+++ bin/pad.rs
@@ -30,6 +30,23 @@ fn language_to_vec_string(language: &str) -> Vec<String> {
     language.split("\n").map(|l| l.to_string()).collect()
 }
 
+fn normalize_vec_string(vec_string: Vec<String>) {
+    let mut composition = String::new();
+    for line in vec_string {
+        let l = line;
+        let copy_l = l.trim_start();
+        if copy_l.starts_with("--") {
+            composition.push_str("\n");
+        //} else if is_import(copy_l.to_string()) {
+        //imports_needed.push(copy_l.to_owned());
+        //composition.push_str("\n");
+        } else {
+            composition.push_str("\n");
+            composition.push_str(&l);
+        }
+    }
+}
+
 #[allow(unused_variables)]
 fn run() -> Result<(), Error> {
     let filename = "songs/test/template.socool";
@@ -42,16 +59,10 @@ fn run() -> Result<(), Error> {
         language.push_str("\n");
     });
 
-    let from_filename = filename_to_vec_string(filename);
-    let from_language = language_to_vec_string(language.as_str());
+    let from_filename = normalize_vec_string(filename_to_vec_string(filename));
+    let from_language = normalize_vec_string(language_to_vec_string(language.as_str()));
 
-    for (a, b) in from_filename.iter().zip(&from_language) {
-        let copy_a = a.trim_start();
-        if copy_a.starts_with("--") {
-            dbg!(1234);
-        };
-        assert_eq!(a, b);
-    }
+    assert_eq!(from_language, from_filename);
 
     Ok(())
 }
