+++ parser/working.socool
@@ -48,6 +48,7 @@ main = {
         AD(1/100, 1/1, 2) |
         Portamento 1/10 |
         Sequence [
+            --AsIs,
             AsIs,
             foo,
             Tm 3/2 | Length 2.0
