+++ src/generation/interpretable.rs
@@ -1,7 +1,4 @@
 use crate::generation::{RenderReturn, RenderType};
-use std::fs::File;
-use std::io::prelude::*;
-use std::io::BufReader;
 
 pub enum InputType<'a> {
     Filename(&'a str),
@@ -16,51 +13,14 @@ impl Interpretable for InputType<'_> {
     fn make(&self, _target: RenderType) -> RenderReturn {
         match &self {
             InputType::Filename(filename) => {
-                let vec_string = filename_to_vec_string(filename);
+                //let vec_string = filename_to_vec_string(filename);
                 unimplemented!();
             }
             InputType::Language(language) => {
-                let vec_string = language_to_vec_string(language);
+                //let vec_string = language_to_vec_string(language);
                 unimplemented!();
             }
         }
     }
 }
 
-fn filename_to_vec_string(filename: &str) -> Vec<String> {
-    let f = File::open(filename).expect("could open file");
-    let file = BufReader::new(&f);
-    file.lines()
-        .map(|line| {
-            let l = line.expect("Could not parse line");
-            l
-        })
-        .collect()
-}
-
-fn language_to_vec_string(language: &str) -> Vec<String> {
-    language.split("\n").map(|l| l.to_string()).collect()
-}
-
-mod tests {
-    #[test]
-    fn import_test() {
-        use super::*;
-        let filename = "songs/test/template.socool";
-        let mut language = "".to_string();
-        let f = File::open(filename).expect("couldn't open song/test/template.socool");
-        let file = BufReader::new(&f);
-        file.lines().for_each(|line| {
-            let l = line.expect("Could not parse line");
-            language.push_str(&l);
-            language.push_str("\n");
-        });
-
-        let from_filename = filename_to_vec_string(filename);
-        let from_language = language_to_vec_string(language.as_str());
-
-        for (a, b) in from_filename.iter().zip(&from_language) {
-            assert_eq!(a, b);
-        }
-    }
-}
